requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appnope
asttokens
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm
dataclasses-json==0.6.7
debugpy
decorator
distro==1.9.0
einops==0.8.1
executing
faiss-cpu==1.7.4
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
idna==3.10
ipykernel
ipython
ipywidgets==8.1.5
jedi
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client
jupyter_core
jupyterlab_widgets==3.0.13
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.41
langchain-experimental==0.3.4
langchain-groq==0.2.4
langchain-huggingface==0.1.2
langchain-tavily==0.1.5
langchain-text-splitters==0.3.6
langgraph==0.3.20
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.4
langgraph-sdk==0.1.58
langsmith==0.3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline
mpmath==1.3.0
multidict==6.1.0
mypy==1.15.0
mypy-extensions==1.0.0
narwhals==1.37.1
nest-asyncio
networkx==3.4.2
numpy==1.26.4
orjson==3.10.15
ormsgpack==1.9.0
packaging
pandas==2.2.3
parso
pexpect
pillow==11.1.0
platformdirs
prompt-toolkit
propcache==0.2.1
protobuf==6.30.2
psutil
ptyprocess
pure-eval
pyarrow==20.0.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments
pypdf==5.3.1
python-dateutil
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
six
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
stack-data
streamlit==1.45.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
tornado
tqdm==4.67.1
traitlets
transformers==4.49.0
typing-inspect==0.9.0
typing_extensions
tzdata==2025.1
urllib3==2.3.0
wcwidth
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
